refactor: Seperate auth conditional directory
