+++ src/app/layout.tsx
@@ -3,9 +3,7 @@ import { Geist, Geist_Mono } from 'next/font/google';
 import './globals.css';
 import ThemeInitializer from '@components/system/ThemeInitializer';
 import { Toaster } from 'react-hot-toast';
-import SideBarWrapper from '@components/layout/SideBarWrapper';
 import { InitUserStore } from '@stores/userStore';
-import RightPanelWrapper from '@components/layout/RightPannelWrapper';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -18,30 +16,24 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: '코테는 정말 귀여워',
+  description: '코딩하기',
 };
 
 export default function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
   return (
     <html lang="ko">
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
-        <div className="flex h-screen">
-          <SideBarWrapper />
-          <div className="flex flex-1 overflow-hidden">
-            <main className="flex-1 overflow-auto">{children}</main>
-            <RightPanelWrapper />
-          </div>
-        </div>
         <InitUserStore />
         <ThemeInitializer />
         <Toaster position="top-center" reverseOrder={false} />
+        {children}
       </body>
     </html>
   );
