+++ src/components/layout/RightPanel/HomePanel.tsx
@@ -5,7 +5,7 @@ import PetStatus from '@components/home/PetStatus';
 
 export default function HomePanel() {
   return (
-    <div className="flex flex-col gap-4 p-4">
+    <div className="flex flex-col gap-4 p-8">
       <LoginStreak />
       <PetStatus />
     </div>
