--- src/components/layout/RightPannelWrapper.tsx
@@ -1,10 +1,0 @@
-import RightPanel from './RightPanel';
-import AuthConditional from '../common/AuthConditional';
-
-const RightPanelWrapper = () => (
-  <AuthConditional>
-    <RightPanel />
-  </AuthConditional>
-);
-
-export default RightPanelWrapper;
