--- src/components/layout/SideBarWrapper.tsx
@@ -1,10 +1,0 @@
-import Sidebar from '@components/sidebar/SideBar';
-import AuthConditional from '@components/common/AuthConditional';
-
-const SideBarWrapper = () => (
-  <AuthConditional>
-    <Sidebar />
-  </AuthConditional>
-);
-
-export default SideBarWrapper;
